assignment 3

+++ NodeJS/conFusionServer/routes/dishRouter.js
@@ -42,29 +42,39 @@ dishRouter.route('/')
 // if authentication fail, passport authenticator 
 // will respond to the client
 .post(authenticate.verifyUser, (req,res,next) => {
-    Dishes.create(req.body)
-    .then((dish) => {
-        console.log("Dish Created ", dish);
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dish);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Dishes.create(req.body)
+        .then((dish) => {
+            console.log("Dish Created ", dish);
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dish);
+        }, (err) => next(err))
+        .catch((err) => next(err));
 })
 // since put operation on /dishes doesn't make sense??
 // maybe it means not possible to edit all dishes at once?
+/*.put(authenticate.verifyUser, (req,res,next) => {
+    res.statusCode = 403;
+    res.end('PUT operation not supported on /dishes');
+})*/
 .put(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.remove({})
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Dishes.remove({})
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
 });
 
 /* for dishes/:dishId */
@@ -80,30 +90,36 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
-    res.statusCode = 403;
-    res.end('POST operation not supported on /dishes/'
-    + req.params.dishId);
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        res.statusCode = 403;
+        res.end('POST operation not supported on /dishes/'
+        + req.params.dishId);
 })
 .put(authenticate.verifyUser, (req,res,next) => {
-    Dishes.findByIdAndUpdate(req.params.dishId, {
-        $set: req.body
-        // will return the updated dish as json string in the reply
-    }, { new: true })
-    .then((dish) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dish);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Dishes.findByIdAndUpdate(req.params.dishId, {
+            $set: req.body
+            // will return the updated dish as json string in the reply
+        }, { new: true })
+        .then((dish) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dish);
+        }, (err) => next(err))
+        .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.findByIdAndRemove(req.params.dishId)
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Dishes.findByIdAndRemove(req.params.dishId)
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
 });
 
 // comments
@@ -164,31 +180,35 @@ dishRouter.route('/:dishId/comments')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req,res,next) => {
-    res.statusCode = 403;
-    res.end('PUT operation not supported on /dishes/' 
-        + req.params.dishId + '/comments');
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        res.statusCode = 403;
+        res.end('PUT operation not supported on /dishes/' 
+            + req.params.dishId + '/comments');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.findById(req.params.dishId)
-    .then((dish) => {
-        if(dish != null) {
-            for(var i = (dish.comments.length - 1); i >= 0; i--){
-                dish.comments.id(dish.comments[i]._id).remove();
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Dishes.findById(req.params.dishId)
+        .then((dish) => {
+            if(dish != null) {
+                for(var i = (dish.comments.length - 1); i >= 0; i--){
+                    dish.comments.id(dish.comments[i]._id).remove();
+                }
+                dish.save()
+                .then((dish) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);
+                }, (err) => next(err));
             }
-            dish.save()
-            .then((dish) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);
-            }, (err) => next(err));
-        }
-        else {
-            err = new Error('Dish ' + req.params.dishId + ' not found');
-            err.status = 404;
-            return next(err); //app.js will handle error
-        }
-    }, (err) => next(err))
-    .catch((err) => next(err));
+            else {
+                err = new Error('Dish ' + req.params.dishId + ' not found');
+                err.status = 404;
+                return next(err); //app.js will handle error
+            }
+        }, (err) => next(err))
+        .catch((err) => next(err));
 });
 
 /* /:dishId/comments/:commentId */
@@ -223,24 +243,32 @@ dishRouter.route('/:dishId/comments/:commentId')
 .put(authenticate.verifyUser, (req,res,next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-        if(dish != null && dish.comments.id(req.params.commentId) != null) {
-            // update the rating or comment only
-            if(req.body.rating) {
-                dish.comments.id(req.params.commentId).rating = req.body.rating;
+        var comment = dish.comments.id(req.params.commentId);
+        if(dish != null && comment != null) {
+            if(req.user._id.equals(comment.author)){
+                // update the rating or comment only
+                if(req.body.rating) {
+                    dish.comments.id(req.params.commentId).rating = req.body.rating;
+                }
+                if(req.body.comment) {
+                    dish.comments.id(req.params.commentId).comment = req.body.comment;
+                }
+                dish.save()
+                .then((dish) => {
+                    Dishes.findById(dish._id) 
+                        .populate('comments.author')
+                        .then((dish) => {
+                            res.statusCode = 200;
+                            res.setHeader('Content-Type', 'application/json');
+                            res.json(dish.comments);
+                        })
+                }, (err) => next(err));
             }
-            if(req.body.comment) {
-                dish.comments.id(req.params.commentId).comment = req.body.comment;
+            else {
+                err = new Error('You do not have the permission to perform this action');
+                err.status = 403;
+                return next(err); 
             }
-            dish.save()
-            .then((dish) => {
-                Dishes.findById(dish._id) 
-                    .populate('comments.author')
-                    .then((dish) => {
-                        res.statusCode = 200;
-                        res.setHeader('Content-Type', 'application/json');
-                        res.json(dish.comments);
-                    })
-            }, (err) => next(err));
         }
         else if (dish == null) {
             err = new Error('Dish ' + req.params.dishId + ' not found');
@@ -258,19 +286,28 @@ dishRouter.route('/:dishId/comments/:commentId')
 .delete(authenticate.verifyUser, (req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-        if(dish != null && dish.comments.id(req.params.commentId) != null) {
-            // delete specific comment
-            dish.comments.id(req.params.commentId).remove();
-            dish.save()
-            .then((dish) => {
-                Dishes.findById(dish_.id)
-                    .populate('comments.author')
-                    .then((dish) => {
-                        res.statusCode = 200;
-                        res.setHeader('Content-Type', 'application/json');
-                        res.json(dish);
-                    })
-            }, (err) => next(err));
+        var comment = dish.comments.id(req.params.commentId);
+        if(dish != null && comment != null) {
+            if(req.user._id.equals(comment.author)){
+                // delete specific comment
+                dish.comments.id(req.params.commentId).remove();
+                dish.save()
+                .then((dish) => {
+                    Dishes.findById(dish._id)
+                        .populate('comments.author')
+                        .then((dish) => {
+                            res.statusCode = 200;
+                            res.setHeader('Content-Type', 'application/json');
+                            res.json(dish);
+                        })
+                }, (err) => next(err))
+                .catch((err) => next(err));
+            }
+            else {
+                err = new Error('You do not have the permission to perform this action');
+                err.status = 403;
+                return next(err); 
+            }
         }
         else if (dish == null) {
             err = new Error('Dish ' + req.params.dishId + ' not found');
