+++ NodeJS/conFusionServer/routes/promoRouter.js
@@ -21,27 +21,33 @@ promoRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
-    Promotions.create(req.body)
-    .then((promotion) => {
-        console.log("Promotion Created ", promotion);
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(promotion);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Promotions.create(req.body)
+        .then((promotion) => {
+            console.log("Promotion Created ", promotion);
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(promotion);
+        }, (err) => next(err))
+        .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req,res,next) => {
-    res.statusCode = 403;
-    res.end('PUT operation not supported on /promotions');
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        res.statusCode = 403;
+        res.end('PUT operation not supported on /promotions');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Promotions.remove({})
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Promotions.remove({})
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
 });
 
 
@@ -56,30 +62,36 @@ promoRouter.route('/:promoId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
-    res.statusCode = 403;
-    res.end('POST operation not supported on /promotions/'
-    + req.params.promoId);
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        res.statusCode = 403;
+        res.end('POST operation not supported on /promotions/'
+        + req.params.promoId);
 })
 .put(authenticate.verifyUser, (req,res,next) => {
-    Promotions.findByIdAndUpdate(req.params.promoId, {
-        $set: req.body
-        // will return the updated promotion as json string in the reply
-    }, { new: true })
-    .then((promotion) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(promotion);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Promotions.findByIdAndUpdate(req.params.promoId, {
+            $set: req.body
+            // will return the updated promotion as json string in the reply
+        }, { new: true })
+        .then((promotion) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(promotion);
+        }, (err) => next(err))
+        .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Promotions.findByIdAndRemove(req.params.promoId)
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    authenticate.verifyAdmin(req,res,next);
+    }, (req, res, next) => { 
+        Promotions.findByIdAndRemove(req.params.promoId)
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
 });
 
 module.exports = promoRouter;
